docs: update Readme, simplify demo file structure
