--- examples/demo/src/app/todos/models/index.ts
@@ -1,4 +1,0 @@
-// start:ng42.barrel
-export * from './todos.model';
-export * from './todos.schema';
-// end:ng42.barrel
