--- examples/demo/src/app/todos/services/index.ts
@@ -1,3 +1,0 @@
-// start:ng42.barrel
-export * from './todos.service';
-// end:ng42.barrel
