--- examples/demo/src/app/todos/todos-routing.module.ts
@@ -1,21 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { TodosComponent } from './components/todos/todos.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    children: [
-      {
-        path: '',
-        component: TodosComponent,
-      },
-    ],
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class TodosRoutingModule {}
